2020/12/18宣传用品管理-印刷品,宣传画面设计 物料制作 导入 [ci skip]

+++ src/main/java/com/hbhb/cw/publicity/model/PrintNotice.java
@@ -21,7 +21,7 @@ public class PrintNotice implements Serializable {
     /**
      * 印刷品id
      */
-    private Integer printId;
+    private Long printId;
     /**
      * 接收人id
      */
